# swagger.yaml
openapi: 3.0.0
info:
  title: DSD Resort API Collection
  version: 1.0.0
  description: Complete OpenAPI Collection for DSD Resort Admin Backend System organized by feature modules.
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
  /api/v1/auth/signup:
    post:
      summary: User Signup
      tags:
        - 1. Authentication Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                role: { type: string }
      responses:
        "200":
          description: Successful response
  /api/v1/auth/login:
    post:
      summary: User Login
      tags:
        - 1. Authentication Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Successful response
...